[BUGFIX] Fix incorrect value passing in from signal to slot
For whatever reason a big university ran into the error "Cannot unpack array with string keys" when exporting courses with the export backend module.
We played a bit and viewed the history of this file and we think the correct change in commit ec6b8df3143deae2af712f3edb26ba192e60797e for the line
`foreach ($record as $propertyName => $property) {`
would have been
`foreach (array_values($record) as $property) {`
In addition we are using a numerized array instead of an associative array when passing variables to the slot.

Related: https://projekte.in2code.de/issues/53322